display files inside folder done

+++ controllers/file_controller.php
@@ -5,20 +5,26 @@ require_once('model/log.php');
 
 function upload_action() {
 	if ($_SERVER["REQUEST_METHOD"] === "POST") {
-		/*
-		if (preg_match('#[\x00-\x1F\x7F-\x9F/\\\\]#', $name_file)) {
-		    exit("Nom de fichier non valide");
-		}
-		*/
-		if (file_check_upload($_FILES)) {
-			file_upload($_FILES);
-			write_log('access.log', 'File uploaded.');
-			header('Location: ?action=home');
-			exit(0);
+		if (!empty($_FILES['userfile']['name'])) {
+			/*
+			if (preg_match('#[\x00-\x1F\x7F-\x9F/\\\\]#', $name_file)) {
+			    exit("Nom de fichier non valide");
+			}
+			*/
+			if (file_check_upload($_FILES)) {
+				file_upload($_FILES);
+				write_log('access.log', 'File ' . $_FILES['userfile']['name'] . ' uploaded.');
+				header('Location: ?action=home');
+				exit(0);
+			}
+			else {
+				write_log('security.log', 'Error upload transfert.');
+				echo "<p style='color:white;font-family:Calibri'>" . "Erreur" . $_FILES['userfile']['error'] . "lors du transfert : Vous n'avez pas sélectionné de fichier, ou celui-ci est trop lourd." . "</p>";
+			}
 		}
 		else {
-			write_log('security.log', 'Error upload transfert.');
-			echo "<p style='color:white;font-family:Calibri'>" . "Erreur" . $_FILES['userfile']['error'] . "lors du transfert : Vous n'avez pas sélectionné de fichier, ou celui-ci est trop lourd." . "</p>";
+			write_log('security.log', 'No file to upload.');
+			echo "<p style='color:white;font-family:Calibri'>" . "Choisissez un fichier." . "</p>";
 		}
 	}
 	require('views/home.html');
@@ -26,15 +32,23 @@ function upload_action() {
 
 function rename_action() {
 	if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-		if (file_check_permission()) {
-			file_rename($_POST);
-			write_log('access.log', 'File renamed.');
-			header('Location: ?action=home');
-			exit(0);
+		$error = '';
+		if (!empty($_POST['filename'])) {
+			if (file_check_permission()) {
+				file_rename($_POST);
+				write_log('access.log', 'File ' . $_POST['input-filename'] . ' renamed.');
+				header('Location: ?action=home');
+				exit(0);
+			}
+			else {
+				write_log('security.log', 'Error file access for rename action.');
+				echo "<p style='color:white;font-family:Calibri'>" . "Vous n'avez pas accès à ce fichier" . "</p>";
+			}
 		}
 		else {
-			write_log('security.log', 'Error file access for rename action.');
-			echo "<p style='color:white;font-family:Calibri'>" . "Vous n'avez pas accès à ce fichier" . "</p>";
+			write_log('security.log', 'No name to rename file.');
+			$error = 'Choisissez un nom.';
+			//echo "<p style='color:white;font-family:Calibri'>" . "Choisissez un nom." . "</p>";
 		}
 	}
 	require('views/home.html');
@@ -42,25 +56,31 @@ function rename_action() {
 
 function replace_action() {
 	if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-		if (file_check_permission()) {
-			file_replace($_POST);
-			write_log('access.log', 'File replaced.');
-			header('Location: ?action=home');
-			exit(0);
+		if (!empty($_FILES['filename']['name'])) {
+			if (file_check_permission()) {
+				file_replace($_POST);
+				write_log('access.log', 'File ' . $_POST['input-filename'] . ' replaced.');
+				header('Location: ?action=home');
+				exit(0);
+			}
+			else {
+				write_log('security.log', 'Error file access for replace action.');
+				echo "<p style='color:white;font-family:Calibri'>" . "Vous n'avez pas accès à ce fichier" . "</p>";
+			}
 		}
 		else {
-			write_log('security.log', 'Error file access for replace action.');
-			echo "<p style='color:white;font-family:Calibri'>" . "Vous n'avez pas accès à ce fichier" . "</p>";
+			write_log('security.log', 'No file to replace file.');
+			echo "<p style='color:white;font-family:Calibri'>" . "Choisissez un fichier." . "</p>";
 		}
 	}
 	require('views/home.html');
 }
 
 function delete_action() {
-	// if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+	if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 		if (file_check_permission()) {
 			file_delete($_POST);
-			write_log('access.log', 'File deleted.');
+			write_log('access.log', 'File ' . $_POST['input-filename'] . ' deleted.');
 			header('Location: ?action=home');
 			exit(0);
 		}
@@ -68,38 +88,50 @@ function delete_action() {
 			write_log('security.log', 'Error file access for delete action.');
 			echo "<p style='color:white;font-family:Calibri'>" . "Vous n'avez pas accès à ce fichier" . "</p>";
 		}
-	// }
+	}
 	require('views/home.html');
 }
 
 function modify_action() {
-	// if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-		if (file_check_permission()) {
-			file_modify($_POST);
-			write_log('access.log', 'File modified.');
-			header('Location: ?action=home');
-			exit(0);
+	if ($_SERVER['REQUEST_METHOD'] === 'POST') { 
+		if (!empty($_POST['content-modification'])) {
+			if (file_check_permission()) {
+				file_modify($_POST);
+				write_log('access.log', 'File ' . $_POST['input-filename'] . ' modified.');
+				header('Location: ?action=home');
+				exit(0);
+			}
+			else {
+				write_log('security.log', 'Error file access for modify action.');
+				echo "<p style='color:white;font-family:Calibri'>" . "Vous n'avez pas accès à ce fichier" . "</p>";
+			}
 		}
 		else {
-			write_log('security.log', 'Error file access for modify action.');
-			echo "<p style='color:white;font-family:Calibri'>" . "Vous n'avez pas accès à ce fichier" . "</p>";
+			write_log('security.log', 'No content to modify file.');
+			echo "<p style='color:white;font-family:Calibri'>" . "Entrez du texte." . "</p>";
 		}
-	// }
+	}
 	require('views/home.html');
 }
 
 function move_action() {
-	// if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-		if (file_check_permission()) {
-			file_move($_POST);
-			write_log('access.log', 'File moved.');
-			header('Location: ?action=home');
-			exit(0);
+	if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+		if (!empty($_POST['new-folder'])) {
+			if (file_check_permission()) {
+				file_move($_POST);
+				write_log('access.log', 'File ' . $_POST['input-filename'] . ' moved.');
+				header('Location: ?action=home');
+				exit(0);
+			}
+			else {
+				write_log('security.log', 'Error file access for move action.');
+				echo "<p style='color:white;font-family:Calibri'>" . "Vous n'avez pas accès à ce fichier" . "</p>";
+			}
 		}
 		else {
-			write_log('security.log', 'Error file access for move action.');
-			echo "<p style='color:white;font-family:Calibri'>" . "Vous n'avez pas accès à ce fichier" . "</p>";
+			write_log('security.log', 'No folder to move file.');
+			echo "<p style='color:white;font-family:Calibri'>" . "Choisissez un nom." . "</p>";
 		}
-	// }
+	}
 	require('views/home.html');
 }
